- Updated package names to better suit java naming convention - Updated dependencies in pom.xml to use spring-boot-starter-web - Added StoreClientInterface - Moved RequestExecutor, StoreClient and StoreClientData to different packages
