+++ pom.xml
@@ -1,25 +1,27 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
+    <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.example</groupId>
-  <artifactId>myPetStoreTests</artifactId>
-  <version>1.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
+    <groupId>org.example</groupId>
+    <artifactId>myPetStoreTests</artifactId>
+    <version>1.0-SNAPSHOT</version>
+    <packaging>jar</packaging>
 
-  <name>myPetStoreTests</name>
-  <url>http://maven.apache.org</url>
+    <name>myPetStoreTests</name>
+    <url>http://maven.apache.org</url>
 
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
-    <junit.jupiter.version>5.9.1</junit.jupiter.version>
-    <lombok.version>1.18.36</lombok.version>
-    <slf4j.version>1.7.36</slf4j.version>
-    <owner.version>1.0.11</owner.version>
-    <spring-boot-starter-web.version>3.5.0</spring-boot-starter-web.version>
-    <httpclient5.version>5.3.1</httpclient5.version>
-  </properties>
+        <junit.jupiter.version>5.9.1</junit.jupiter.version>
+        <lombok.version>1.18.36</lombok.version>
+        <slf4j.version>1.7.36</slf4j.version>
+        <owner.version>1.0.11</owner.version>
+        <spring-boot-starter-web.version>3.5.0</spring-boot-starter-web.version>
+        <httpclient5.version>5.3.1</httpclient5.version>
+    </properties>
 
   <dependencies>
     <dependency>
